@@ -1,6 +1,6 @@
 {
   "name": "react-js-carplay",
-  "version": "2.0.2",
+  "version": "2.1.0",
   "private": false,
   "main": "dist/Carplay.js",
   "module": "dist/Carplay.js",
@@ -27,11 +27,13 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "jmuxer": "^2.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-modal": "^3.14.4",
-    "react-scripts": "4.0.3",
+    "jmuxer": "^2.0.5",
+    "react": "^18.2.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^18.2.0",
+    "react-modal": "^3.16.1",
+    "react-scripts": "^5.0.1",
+    "stream-browserify": "^3.0.0",
     "web-vitals": "^1.1.2"
   },
   "babel": {
@@ -40,9 +42,9 @@
     ]
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "publish:npm": "rm -rf dist && mkdir dist && babel ./src -d dist --copy-files"
   },
